pip-requirements.txt
Django==1.9
MySQL-python==1.2.5
Pillow==3.1.1
argparse==1.2.1
colorama==0.3.7
django-debug-toolbar==1.4
numpy==1.10.4
python-dateutil==2.5.1
qrcode==5.2.2
simplejson==3.8.2
six==1.10.0
sqlparse==0.1.19
wsgiref==0.1.2
